pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lightweight-local-assistant"
version = "0.1.0"
description = "A powerful, lightweight Python library for local AI automation and secure bridge between Gemini CLI and local AI models."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Duncan MacIntyre", email = "duncanmacintyre@users.noreply.github.com" }
]
dependencies = [
    "ollama>=0.4.7",
    "python-dotenv>=1.0.1",
    "pypdf>=5.2.0",
    "mcp>=1.3.0",
    "fastmcp>=0.4.1",
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "httpx>=0.28.1"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["lightweight_local_assistant*"]
